Move xs1 base entity to separate module
