+++ homeassistant/components/xs1/switch.py
@@ -11,7 +11,8 @@ from homeassistant.core import HomeAssistant
 from homeassistant.helpers.entity_platform import AddEntitiesCallback
 from homeassistant.helpers.typing import ConfigType, DiscoveryInfoType
 
-from . import ACTUATORS, DOMAIN as COMPONENT_DOMAIN, XS1DeviceEntity
+from . import ACTUATORS, DOMAIN as COMPONENT_DOMAIN
+from .entity import XS1DeviceEntity
 
 
 def setup_platform(
